[testing] add a test runner for iPhone 8, iOS 13. (#8220)
Part of b/137114947.

This PR added a new test runner for iOS 13 tests and its supporting functions in bzl.

Because of the limitation of using `select` on `runners` (which is not a rule attribute), we can't directly replace a list of runner with another list of runners. In addition, `select`'s default value doesn't support `None`. So, we pulled a runner (iPhoneX) from the default runner and used it in a separate unit test together with an iOS 13 runner. 

The only difference is the name of this specific test is changed to `unit_test_extension`.