[Banner] Make default textView background color clear (#8258)
When a client sets the background color of a banner it can become out of sync with the textView's background color. Currently the textView's background color defaults to `UIColor.whiteColor` so it is not a problem but if a client changes the background color as in the example below then the textView's background is shown.

| Before | After |
| --- | --- |
|<img width="415" alt="Screen Shot 2019-08-09 at 9 51 32 AM" src="https://user-images.githubusercontent.com/7131294/62784044-93899a80-ba8b-11e9-8c0a-dae0d07be3a3.png">|<img width="417" alt="Screen Shot 2019-08-09 at 9 52 01 AM" src="https://user-images.githubusercontent.com/7131294/62784055-97b5b800-ba8b-11e9-8a46-5aaeb5e92e30.png">|

Related to b/139125319