[Catalog] Remove a layer of indirection in AppTheme. (#8261)
AppTheme was originally created to store each of the Scheme types (Typography, Color, etc...). We've since introduced a container scheme type which can be used for the same purpose, so this PR removes the overlapping functionality in favor of simply exposing a global static container scheme on the AppTheme type.

Closes https://github.com/material-components/material-components-ios/issues/8260