[ActionSheet] Add dark mode support to theming extension (#8245)
This PR adds dark mode support to the theming extension for MDCActionSheetController. In order to support dark mode clients will need to set dynamic colors in their color scheme. 

There is no change to `mdc_elevationDidChangeBlock` because the elevation on an MDCActionSheetController should never change since it is not a settable property. If we later expose a setter then we will need to add that functionality. Additionally unit test were added but since MDCActionSheetController doesn't allow for subclassing I could not test for traitCollection changes. In order to test this more thoroughly do the steps as follows.

1. Launch MDCCatalog on an iOS 13 simulator.
2. Open the ActionSheet example.
3. Switch between light and dark mode in Xcode 11.

Following the steps above, this should be the result.
![actionSheet](https://user-images.githubusercontent.com/7131294/62724608-61bbf980-b9e1-11e9-8979-d586b49d2c58.gif)