[Shapes] Update border color to support dynamic color (#8308)
### Context

Currently the `setBorderColor` method of MDCShapedShadowLayer can receive a dynamic color but will use the light mode color because it doesn't not know about traitCollections since it is a CALayer subclass. This causes a problem for any component that subclasses `UIControl` and updates its visuals based on state. For instance MDCCard within `setHighlighted` calls `updateBorderColor` which then would call `setBorderColor` on the shapedShadowLayer. Since traitCollection was never involved here the border gets render as it's in a light mode even if it happens to be in a dark mode. Additionally, we may need this for other CGColor things but this will unblock card for now.

### Screenshots

| Before | After |
| - | - |
|![before](https://user-images.githubusercontent.com/7131294/62970425-73d0da00-bddd-11e9-8643-6ae5b039a3a2.gif)|![after](https://user-images.githubusercontent.com/7131294/62970438-7d5a4200-bddd-11e9-9a38-45b7e8bb8328.gif)|