[NavigationDrawer] Respect the rotation preferences of presenting controller (#8278)
This pull request changes the behaviour of the NavigationDrawer, so it respects the rotation preference of the presentating controller. Currently, if the NavigationDrawer is presented, rotation of the device can change the orientation of the NavigationDrawer and the presentation controller, even if the presentating controller does not support it. Dismissing the NavigationDrawer in the rotated position will not revert the orientation of the presenting controller. This may result in unexpected outcomes for the presenting controller. This behaviour is not observed in other components which pop onto the screen (e.g. MDCBottomSheetController will not override the rotation preference of the presenting controller)