[Ripple] Performance Improvements on initialization (#8354)
To improve performance when Ripple is being initialized, I have tried to look at what things are not needed, or can be delayed to a later stage, like the ripple invocation and not init.

self.backgroundColor is initially nil, which defaults to clearColor, hence we can remove that setting.

self.layer.maskToBounds = YES is redundant because maskToBounds is dependent on the rippleStyle which is set anyway in the method updateRippleStyle that is invoked.

updateRippleStyle does not need to be called in every layout of the view/subviews, because the masking of the ripple based on the shadowPath (if exists) is only important if there is an actual ripple layer being created, otherwise it is masking the "empty case"

The initialization of the maskLayer and setting the delegate only needs to be set when the maskLayer is actually used, so it is then set up if needed only if the ripple is bounded and there is a shadowPath, and the maskLayer is going to be used.

I've run instrumentation of Counters using an iPhone 5c 8GB which uses an A6 chip.
During instrumentation on high frequency, I had gathered that updateRippleStyle takes 0.1ms, and setting backgroundColor 0.2ms every time.