[NavDrawer] Allow animations alongside movement in MDCBottomNavigatio… (#8474)
Add delegate methods to allow the presenting VC to gain access to the transitionCoordinator used during the present/dismiss transitions. Also provide callbacks for y-offset changes so that the presenting VC can animate content along with the y-offset changes.

Fixed requests from previous PR and added some tests for the callbacks.

Screen recording of the usage:

[navdrawer.mov.zip](https://github.com/material-components/material-components-ios/files/3633638/navdrawer.mov.zip)