[NavigationDrawer] Allow header to expand even if content doesn't fill screen. (#8484)
Previously the MDCBottomDrawerViewController would not expand the header if the content did not reach full screen. This change makes it so that clients can have the header always expand the header regardless of the total content height. This new behavior is behind a behavioral flag so that clients can opt into this if they would prefer that. This change is so that clients can still have their content visible when the header goes into the safe area.

## Screenshots
_iPhone XS Max_

| Before | After |
| - | - |
|![beforeXSMax](https://user-images.githubusercontent.com/7131294/65453465-15dddc00-ddf8-11e9-9a72-01368a285533.gif)|![largeDeviceAfter](https://user-images.githubusercontent.com/7131294/65453424-052d6600-ddf8-11e9-9d7d-efaa4033cd46.gif)|

_iPhone 4s_

| Before | After |
| - | - |
|![4sbefore](https://user-images.githubusercontent.com/7131294/65453537-458ce400-ddf8-11e9-9725-3a4fa843d58d.gif)|![4safter](https://user-images.githubusercontent.com/7131294/65453548-4d4c8880-ddf8-11e9-9790-28f9fa32142d.gif)|

Closes #7684