[TextFields] Add some important protocols (#8493)
This PR adds the MDCContainedInputView protocol and the MDCContainedInputViewStyle protocol. MDCContainedInputView is the protocol that all the contained input views conform to. It has properties for all of the things that contained input views should have in common. This includes a style property, which is an object conforming to MDCContainedInputViewStyle. The main job of the style object is to apply various visual decorations on the textfield, like filled backgrounds for filled textfields or outlines for outlined textfields. The style objects also vend the vertical positioning reference objects, which are the things that determine the vertical distances between the textfield's subviews. This makes sense when you consider that the differences in those distances depend on the style of the textfield, i.e. whether it's filled or outlined.

Related to #6942.