[TextFields] Rename "contained input view" to "text control" (#8524)
In this PR I am beginning the naming transition from "contained input view" to "text control." This is just a renaming PR, there are no other changes.

At some point after this I want to make it its own component, rather than an extension on TextFields, in the spirit of Ripple and Ink.

Related to #6942.